paperbanana-service/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
python-multipart>=0.0.12
boto3>=1.35.0
pillow>=10.0
httpx>=0.27.0
pydantic>=2.0
python-dotenv>=1.0
aiofiles>=23.0
google-generativeai>=0.8.0
